More consistency fixes (#35)
* Consistency fixes like strict types & float-int fixes.

* Apply fixes from StyleCI (#36)

+++ database/migrations/2018_06_07_123211_plans.php
@@ -69,7 +69,7 @@ class Plans extends Migration
             $table->integer('subscription_id');
 
             $table->string('code');
-            $table->integer('used')->default(0);
+            $table->float('used', 9, 2)->default(0);
 
             $table->timestamps();
         });
